Move encoding_util to gui/base.
encoding_util is used only by mozc_tool.  Moving it to src/gui/base
consolidates build dependencies and improves code portability.

This is just a refactoring.  No behavior change is intended.

BUG=#252
TEST=unittest
REF_BUG=19010851
REF_CL=87592180
