+++ src/gyp/tests.gyp
@@ -42,8 +42,7 @@
         '../dictionary/file/dictionary_file_test.gyp:dictionary_file_all_test',
         '../dictionary/system/system_dictionary_test.gyp:system_dictionary_all_test',
         '../handwriting/handwriting_test.gyp:handwriting_all_test',
-        # Currently 'gui_all_test' does not exist.
-        # '../gui/gui.gyp:gui_all_test',
+        '../gui/gui.gyp:gui_all_test',
         '../ipc/ipc.gyp:ipc_all_test',
         '../net/net_test.gyp:net_all_test',
         '../prediction/prediction_test.gyp:prediction_all_test',
